requirements.txt
streamlit
pytesseract
transformers
Pillow
torch
